requirements.txt
bson==0.5.10
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==1.0.0
pymongo==4.4.0